Fix compiler warnings and enable debug logging on Windows.
git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@9306 7a7537e8-13f0-0310-91df-b6672ffda945

+++ cups/page.c
@@ -136,10 +136,10 @@ ppdPageSize(ppd_file_t *ppd,		/* I - PPD file record */
       if ((coption = ppdFindCustomOption(ppd, "PageSize")) != NULL)
       {
         if ((cparam = ppdFindCustomParam(coption, "Width")) != NULL)
-	  cparam->current.custom_points = w;
+	  cparam->current.custom_points = (float)w;
 
         if ((cparam = ppdFindCustomParam(coption, "Height")) != NULL)
-	  cparam->current.custom_points = l;
+	  cparam->current.custom_points = (float)l;
       }
 
      /*
